Добавляет отдельный модуль с константами
