+++ src/const.js
@@ -1,0 +1,9 @@
+const SortType = {
+  DEFAULT: 'default',
+  DATE_UP: 'date-up',
+  DATE_DOWN: 'date-down',
+  RATING_UP: 'rating-up',
+  RATING_DOWN: 'rating-down',
+};
+
+export { SortType };
