+++ src/view/sort.js
@@ -1,12 +1,5 @@
 import AbstractView from './abstract';
-
-const SortType = {
-  DEFAULT: 'default',
-  DATE_UP: 'date-up',
-  DATE_DOWN: 'date-down',
-  RATING_UP: 'rating-up',
-  RATING_DOWN: 'rating-down',
-};
+import { SortType } from '../const';
 
 class Sort extends AbstractView {
   constructor() {
